refactor(buildsource): split template_graph.py's graph-folding half, close ADR-061 Phase 5 item 2 in full
template_graph.py's one remaining buildsource/source_graph.py facade
dependency (augment_graph_with_templates reaching _decl_node_id/
_symbol_node_id/_type_node_id) couldn't be fixed in place: the file was
exactly at the repository's 2000-line hard cap. The module's own
docstring already described a parse-vs-fold split (mirroring
type_graph.py/call_graph.py); the graph-folding half moves into a new
sibling module, template_graph_fold.py, whose facade import now reaches
model/graph_identity.py and model/source_graph.py directly. This closes
the last of Phase 5 item 2's four internal-caller exceptions, so item 2
(and Phase 5 as a whole) is now fully D8-compliant.
