+++ tests/test_template_graph.py
@@ -32,16 +32,17 @@ import pytest
 from abicheck.buildsource.build_evidence import BuildEvidence, CompileUnit
 from abicheck.buildsource.source_graph import GraphNode, SourceGraphSummary
 from abicheck.buildsource.template_graph import (
-    EDGE_DECL_INSTANTIATES_TEMPLATE,
-    EDGE_INSTANTIATION_EMITS_SYMBOL,
-    EDGE_TEMPLATE_USES_TYPE,
-    NODE_TEMPLATE_DECL,
-    NODE_TEMPLATE_INSTANTIATION,
     ClangTemplateGraphExtractor,
     TemplateArgUse,
     TemplateInstantiation,
-    augment_graph_with_templates,
     parse_clang_ast_templates,
+)
+from abicheck.buildsource.template_graph_fold import (
+    EDGE_DECL_INSTANTIATES_TEMPLATE,
+    EDGE_INSTANTIATION_EMITS_SYMBOL,
+    NODE_TEMPLATE_DECL,
+    NODE_TEMPLATE_INSTANTIATION,
+    augment_graph_with_templates,
     template_decl_node_id,
     template_instantiation_node_id,
 )
@@ -1751,127 +1752,6 @@ def test_namespace_scoped_explicit_instantiation_detached_outside_namespace_stil
     augment_graph_with_templates(graph, out)
     tdecl_nodes = {n.id for n in graph.nodes if n.kind == NODE_TEMPLATE_DECL}
     assert len([n for n in tdecl_nodes if "apply" in n]) == 1  # one shared node
-
-
-# ── graph augmentation tests ─────────────────────────────────────────────────
-
-
-def test_augment_graph_creates_instantiation_and_template_decl_nodes() -> None:
-    graph = SourceGraphSummary()
-    graph.add_node(
-        GraphNode(
-            id="binary_symbol://_ZNK7WrapperIN8internal6DetailEE3getEv",
-            kind="binary_symbol",
-            label="_ZNK7WrapperIN8internal6DetailEE3getEv",
-        )
-    )
-    inst = TemplateInstantiation(
-        kind="record",
-        template_qname="Wrapper",
-        label="Wrapper<internal::Detail>",
-        args=(TemplateArgUse("internal::Detail", "internal::Detail"),),
-        emitted_symbols=("_ZNK7WrapperIN8internal6DetailEE3getEv",),
-    )
-    added = augment_graph_with_templates(graph, [inst])
-    assert (
-        added == 3
-    )  # DECL_INSTANTIATES_TEMPLATE + TEMPLATE_USES_TYPE + INSTANTIATION_EMITS_SYMBOL
-
-    node_ids = {n.id: n for n in graph.nodes}
-    tdecl_id = template_decl_node_id("Wrapper")
-    tinst_id = template_instantiation_node_id("Wrapper<internal::Detail>")
-    assert node_ids[tdecl_id].kind == NODE_TEMPLATE_DECL
-    assert node_ids[tinst_id].kind == NODE_TEMPLATE_INSTANTIATION
-
-    edge_kinds = {(e.src, e.dst, e.kind) for e in graph.edges}
-    assert (tinst_id, tdecl_id, EDGE_DECL_INSTANTIATES_TEMPLATE) in edge_kinds
-    assert (
-        tinst_id,
-        "type://internal::Detail",
-        EDGE_TEMPLATE_USES_TYPE,
-    ) in edge_kinds
-    assert (
-        tinst_id,
-        "binary_symbol://_ZNK7WrapperIN8internal6DetailEE3getEv",
-        EDGE_INSTANTIATION_EMITS_SYMBOL,
-    ) in edge_kinds
-
-
-def test_augment_graph_enum_argument_mints_enum_type_node_not_record_type() -> None:
-    """``TemplateArgUse.target_decl_kind`` (threaded through by
-    ``_resolve_arg_targets``) must actually be consulted -- an enum template
-    argument should mint an ``enum_type`` node, not the ``record_type``
-    fallback every resolved argument used to get unconditionally (Codex
-    review: ``_type_node_kind`` was computed but never called)."""
-    graph = SourceGraphSummary()
-    inst = TemplateInstantiation(
-        kind="record",
-        template_qname="Wrapper",
-        label="Wrapper<internal::Color>",
-        args=(TemplateArgUse("internal::Color", "internal::Color", "EnumDecl"),),
-    )
-    augment_graph_with_templates(graph, [inst])
-    node_ids = {n.id: n for n in graph.nodes}
-    assert node_ids["type://internal::Color"].kind == "enum_type"
-
-
-def test_augment_graph_typedef_argument_mints_typedef_node() -> None:
-    graph = SourceGraphSummary()
-    inst = TemplateInstantiation(
-        kind="record",
-        template_qname="Wrapper",
-        label="Wrapper<internal::Handle>",
-        args=(TemplateArgUse("internal::Handle", "internal::Handle", "TypedefDecl"),),
-    )
-    augment_graph_with_templates(graph, [inst])
-    node_ids = {n.id: n for n in graph.nodes}
-    assert node_ids["type://internal::Handle"].kind == "typedef"
-
-
-def test_augment_graph_skips_symbol_edge_when_not_exported() -> None:
-    """An instantiated member the graph carries no binary_symbol node for
-    (never ODR-used / inlined away / no binary evidence loaded) gets no
-    INSTANTIATION_EMITS_SYMBOL edge -- ADR-057 D1's join-by-shared-node-id
-    rule, reapplied."""
-    graph = SourceGraphSummary()
-    inst = TemplateInstantiation(
-        kind="function",
-        template_qname="identity",
-        label="identity<int>",
-        emitted_symbols=("_Z8identityIiET_S0_",),
-    )
-    augment_graph_with_templates(graph, [inst])
-    assert not any(e.kind == EDGE_INSTANTIATION_EMITS_SYMBOL for e in graph.edges)
-
-
-def test_augment_graph_unresolved_argument_gets_no_uses_type_edge() -> None:
-    graph = SourceGraphSummary()
-    inst = TemplateInstantiation(
-        kind="record",
-        template_qname="Wrapper",
-        label="Wrapper<int>",
-        args=(TemplateArgUse("int", None),),
-    )
-    augment_graph_with_templates(graph, [inst])
-    assert not any(e.kind == EDGE_TEMPLATE_USES_TYPE for e in graph.edges)
-
-
-def test_augment_graph_two_instantiations_share_one_template_decl_node() -> None:
-    graph = SourceGraphSummary()
-    a = TemplateInstantiation(
-        kind="record", template_qname="Wrapper", label="Wrapper<int>"
-    )
-    b = TemplateInstantiation(
-        kind="record", template_qname="Wrapper", label="Wrapper<internal::Detail>"
-    )
-    augment_graph_with_templates(graph, [a, b])
-    tdecl_nodes = [n for n in graph.nodes if n.kind == NODE_TEMPLATE_DECL]
-    assert len(tdecl_nodes) == 1
-    edges_to_tdecl = [
-        e for e in graph.edges if e.kind == EDGE_DECL_INSTANTIATES_TEMPLATE
-    ]
-    assert len(edges_to_tdecl) == 2
-    assert {e.dst for e in edges_to_tdecl} == {tdecl_nodes[0].id}
 
 
 # ── cross-TU extraction merge ────────────────────────────────────────────────
